[hap2] Split haplib.py into haplib.py and hap.py
hap.py includes more basic components that are used from all hap
modules. haplib.py provides more framework-like components and
shall not be used from transporter and the sub classes.
