+++ server/hap2/hatohol/Makefile.am
@@ -1,5 +1,6 @@
 noinst_SCRIPTS = \
 	__init__.py \
+	hap.py \
 	haplib.py \
 	transporter.py \
 	standardhap.py \
