+++ server/hap2/hatohol/hap.py
@@ -1,0 +1,35 @@
+#!/usr/bin/env python
+# coding: UTF-8
+#
+#  Copyright (C) 2015 Project Hatohol
+#
+#  This file is part of Hatohol.
+#
+#  Hatohol is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU Lesser General Public License, version 3
+#  as published by the Free Software Foundation.
+#
+#  Hatohol is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+#  GNU Lesser General Public License for more details.
+#
+#  You should have received a copy of the GNU Lesser General Public
+#  License along with Hatohol. If not, see
+#  <http://www.gnu.org/licenses/>.
+
+"""
+This module contains basic functions and class for HAP (Hatohol Arm Plugins)
+and is supposed to be used both from haplib and transporter.
+transporter and the sub class shall not import haplib. So the functions and
+classes used in them have to be in this module.
+"""
+
+class Signal:
+    """
+    This class is supposed to raise as an exception in order to
+    propagate some events and jump over stack frames.
+    """
+
+    def __init__(self, restart=False):
+        self.restart = restart
