+++ server/hap2/hatohol/test/TestHap.py
@@ -1,0 +1,34 @@
+#!/usr/bin/env python
+# coding: UTF-8
+"""
+  Copyright (C) 2015 Project Hatohol
+
+  This file is part of Hatohol.
+
+  Hatohol is free software: you can redistribute it and/or modify
+  it under the terms of the GNU Lesser General Public License, version 3
+  as published by the Free Software Foundation.
+
+  Hatohol is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+  GNU Lesser General Public License for more details.
+
+  You should have received a copy of the GNU Lesser General Public
+  License along with Hatohol. If not, see
+  <http://www.gnu.org/licenses/>.
+"""
+
+import unittest
+import hap
+
+class Signal(unittest.TestCase):
+    def test_default(self):
+        obj = hap.Signal()
+        self.assertEquals(False, obj.restart)
+
+    def test_restart_is_true(self):
+        obj = hap.Signal(restart=True)
+        self.assertEquals(True, obj.restart)
+
+
