+++ server/hap2/setup_common.py
@@ -5,5 +5,5 @@ setup(name='hatohol',
       author='Procect Hatohol',
       author_email='hatohol-users@list.sourceforge.net',
       url='https://github.com/project-hatohol/hatohol',
-      py_modules=['hatohol.haplib', 'hatohol.transporter', 'hatohol.standardhap']
+      py_modules=['hatohol.hap', 'hatohol.haplib', 'hatohol.transporter', 'hatohol.standardhap']
      )
